requirements.txt
httpx[http2]
gspread
google-auth
python-dotenv
